fix location encoding and add most save functionality

+++ untitled2/src/Line.java
@@ -1,7 +1,7 @@
 import java.awt.*;
 
 public class Line extends Shapes {
-    public Line(double x1, double y1, double x2, double y2) {
+    public Line(float x1, float y1, float x2, float y2) {
         super.type = 1;
         super.x[0] = x1;
         super.y[0] = y1;
