+++ untitled2/src/Plot.java
@@ -1,7 +1,7 @@
 import java.awt.*;
 
 public class Plot extends Shapes {
-    public Plot(double x, double y) {
+    public Plot(float x, float y) {
         super.type = 0;
         super.x[0] = x;
         super.y[0] = y;
