+++ untitled2/src/Shapes.java
@@ -2,8 +2,8 @@ import java.awt.*;
 
 public class Shapes {
     int type;
-    double[] x = new double[3];
-    double[] y = new double[3];
+    float[] x = new float[3];
+    float[] y = new float[3];
     boolean fill;
 
     Color colourFill;
